% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAvailability.R
\name{checkAvailability}
\alias{checkAvailability}
\title{checkAvailability}
\usage{
checkAvailability(fail = TRUE)
}
\arguments{
\item{fail}{A Boolean (default true) whether the program exits if the command is not available}
}
\value{
A Boolean whether the command is available on the system.
}
\description{
Checks that the needed commands are installed on the system
    Should be executed on start up in an application where ldap is critical
}
\examples{
checkAvailability()
}
