% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginLdap.R
\name{loginLdap}
\alias{loginLdap}
\title{loginLdap}
\usage{
loginLdap(user, pass)
}
\arguments{
\item{user}{a string describing the user's name}

\item{pass}{a string describing the user's password}
}
\value{
success (Boolean), msg (string, error explanation), data (fields about the user)
}
\description{
Logs to user in against the ldap server
}
\examples{
result <- loginLdap("user", "pass")
}
