% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseResult.R
\name{parseResult}
\alias{parseResult}
\title{parseResult}
\usage{
parseResult(blob)
}
\arguments{
\item{blob}{The multiline text returned by the ldap command}
}
\value{
A dictionary with the fields available in the blob.
}
\description{
Parses the string returned by the ldap command to extract fields.
}
\examples{
result <- loginLdap("user", "pass")
    fields <- parseResult(result$data)
}
