% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldapAuthenticationConfig.R
\name{setLdapDomain}
\alias{setLdapDomain}
\title{setLdapServerUrl}
\usage{
setLdapDomain(domain)
}
\arguments{
\item{domain}{The ldap server domain}
}
\value{
Nothing
}
\description{
Sets the ldap server domain
}
\examples{
setLdapDomain('ou=users,dc=example,dc=org')
}
