% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldapAuthenticationConfig.R
\name{setLdapServerUrl}
\alias{setLdapServerUrl}
\title{setLdapServerUrl}
\usage{
setLdapServerUrl(url)
}
\arguments{
\item{url}{The ldap server url}
}
\value{
Nothing
}
\description{
Sets the ldap server url
}
\examples{
setLdapServerUrl('ldap://localhost:1389')
}
